+++ changelog.md
@@ -1,3 +1,11 @@
+## v1.2 [ Release | 20240506 ]
+- Added Support for Unified Hosts + fakenews + gambling.
+### Below are the currently supported hosts:
+- Unified Hosts.
+- Unified hosts + fakenews + gambling + social.
+- Unified hosts + fakenews + gambling + porn + social.
+- more will be added soon...
+
 ## v1.1 [ Release | 20240506 ]
 - Simplified the main script logic.
 - Made a separate dir for the host installer scripts.